requirements.txt
streamlit
tensorflow-cpu
numpy
Pillow
pandas
gdown